Rename FioRequestConfirmationScene2 to FioRequestConfirmationScene
